requirements.txt
# Core dependencies
Pillow>=10.0.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
